Fix handling of clusterTags and SparkVersion in Q/P Tools

Fixes #985

- Refactors the code so that both Q/P tools can handle clusterTags
  correctly
- there is a new property isDatabricks that will be true if the
  eventlogs has clusterTags values
- This fixes a bug where the tools were not able to set the SparkVersion
  correctly because logStart event did not exist in DB events.
